Simplified the process to add messages to the error_message dict - updated tests
Still a WIP
